+++ ase/test/calculator/turbomole/test_turbomole_H2.py
@@ -1,4 +1,5 @@
 # type: ignore
+import numpy as np
 from ase import Atoms
 from ase.calculators.turbomole import Turbomole
 import os.path
@@ -10,14 +11,14 @@ def atoms():
     return Atoms('H2', positions=[(0, 0, 0), (0, 0, 1.1)])
 
 
-def test_turbomole_H2(atoms):
+def test_turbomole_H2_rhf_singlet(atoms):
     # Write all commands for the define command in a string
     define_str = '\n\na coord\n*\nno\nb all sto-3g hondo\n*\neht\n\n\n\n*'
 
     atoms.calc = Turbomole(define_str=define_str)
 
     # Run turbomole
-    atoms.get_potential_energy()
+    assert np.isclose(atoms.get_potential_energy(), -28.205659, atol=1e-5)
 
 
 def test_turbomole_H2_uhf_singlet(atoms):
@@ -26,7 +27,7 @@ def test_turbomole_H2_uhf_singlet(atoms):
     })
 
     # Run turbomole
-    atoms.get_potential_energy()
+    assert np.isclose(atoms.get_potential_energy(), -30.828865, atol=1e-5)
 
     # check that it performed a DFT calculation (i.e. basic inputs were most
     # likely understood, cf. issue #735)